feat: add ImageGenerator class to encapsulate image generation logic
